requirements.txt
telethon>=1.29.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
requests>=2.28.0
beautifulsoup4>=4.11.0
yandex-music>=2.1.0
mutagen>=1.47.0
